rename server action file
