+++ web/client-v2/src/features/favorites/components/List/FavoriteArticleFolderList/actGetFavoriteArticleFolderListQuery.ts
