+++ web/client-v2/src/features/favorites/components/Template/FavoriteArticleFolderListTemplate/actGetFavoriteArticleFolderListTemplateQuery.ts
