+++ web/client-v2/src/features/trends/components/List/TrendArticleList/actGetTrendArticleListQuery.ts
