+++ web/client-v2/src/features/trends/components/Template/TrendArticleDashboardTemplate/actGetTrendArticleDashboardTemplateQuery.ts
